refactor: getWorldContracts
